Confirmation

+++ src/components/list/ListForm.vue
@@ -118,6 +118,7 @@ import CheckboxCustom from '@/components/formElements/CheckboxCustom.vue';
 import ButtonText from '@/components/formElements/ButtonText.vue';
 import ButtonSign from '@/components/formElements/ButtonSign.vue';
 import ErrorMessage from '@/components/textElements/ErrorMessage.vue';
+import { isConfirmed } from '@/settings/confirmationPromise';
 import { List } from '@/models/models';
 import { mapActions, mapGetters } from 'vuex';
 
@@ -249,18 +250,23 @@ export default {
           });
       }
     },
-    deleteList() {
-      this.isRequestProcessing = true;
-      this._deleteList(this.list.id)
-        .then(() => {
-          this.closeListModal();
-        })
-        .catch(error => {
-          this.errorMessage = error.response.data.message;
-        })
-        .finally(() => {
-          this.isRequestProcessing = false;
-        });
+    async deleteList() {
+      const confirmationModalTitle = `are you sure you want to delete list  
+        '${this.edittingListObj?.title}' ?`;
+
+      if (await isConfirmed(confirmationModalTitle)) {
+        this.isRequestProcessing = true;
+        this._deleteList(this.list.id)
+          .then(() => {
+            this.closeListModal();
+          })
+          .catch(error => {
+            this.errorMessage = error.response.data.message;
+          })
+          .finally(() => {
+            this.isRequestProcessing = false;
+          });
+      }
     },
   },
 };
